+++ src/App.tsx
@@ -8,6 +8,8 @@ import CoursesPage from "./pages/public/CoursesPage";
 import TeachersPage from "./pages/public/TeachersPage";
 import MethodologyPage from "./pages/public/MethodologyPage";
 import { ContactPage } from "@mui/icons-material";
+import OnlyRegistered from "./components/protection/OnlyRegistered";
+import ProfilePage from "./pages/private/ProfilePage";
 
 function App() {
 
@@ -15,6 +17,7 @@ function App() {
     <>
       <ResponsiveAppBar />
       <Routes>
+        {/* PUBLIC ROUTES */}
         <Route path="/" element={<MainPage />} />
         <Route path="/login" element={<LoginPage />} />
         <Route path="/signup" element={<SignupPage />} />
@@ -22,6 +25,9 @@ function App() {
         <Route path="/teachers" element={<TeachersPage />} />
         <Route path="/methodology" element={<MethodologyPage />} />
         <Route path="/contact" element={<ContactPage />} />
+        
+        {/* PRIVATE ROUTES */}
+        <Route path="/profile" element={<OnlyRegistered><ProfilePage /></OnlyRegistered>} />
 
 
         <Route path="*" element={<NotFoundPage />} />
